refactor: move validate button palette to ui layer
