+++ packages/ui-primitives/src/index.ts
@@ -4,3 +4,4 @@ export * from './views';
 export * from './TaskInstructionPanel';
 export * from './TaskTileSection';
 export * from './ValidateButton';
+export * from './validateButtonPalette';
